+++ ydb/core/client/server/msgbus_server_pq_metacache.cpp
@@ -12,7 +12,7 @@
 #include <ydb/core/base/counters.h>
 #include <ydb/core/kqp/common/kqp.h>
 #include <ydb/core/base/appdata.h>
-#include <ydb/core/persqueue/pq_database.h>
+#include <ydb/core/persqueue/public/pq_database.h>
 #include <ydb/core/persqueue/cluster_tracker.h>
 
 #include <ydb/library/actors/protos/actors.pb.h>
