+++ ydb/services/deprecated/persqueue_v0/grpc_pq_read_actor.cpp
@@ -4,7 +4,7 @@
 #include <ydb/core/client/server/msgbus_server_persqueue.h>
 #include <ydb/library/services/services.pb.h>
 #include <ydb/core/persqueue/percentile_counter.h>
-#include <ydb/core/persqueue/pq_database.h>
+#include <ydb/core/persqueue/public/pq_database.h>
 #include <ydb/core/persqueue/write_meta.h>
 #include <ydb/core/persqueue/writer/source_id_encoding.h>
 #include <ydb/library/persqueue/topic_parser/type_definitions.h>
