+++ ydb/services/persqueue_v1/actors/write_session_actor.cpp
@@ -8,7 +8,7 @@
 #include <ydb/library/persqueue/topic_parser/counters.h>
 #include <ydb/library/wilson_ids/wilson.h>
 #include <ydb/core/base/wilson_tracing_control.h>
-#include <ydb/core/persqueue/pq_database.h>
+#include <ydb/core/persqueue/public/pq_database.h>
 #include <ydb/core/persqueue/write_meta.h>
 #include <ydb/core/base/feature_flags.h>
 #include <ydb/library/services/services.pb.h>
